Can change colours and images fully

+++ app/src/main/java/com/example/left4candy/placeholderapp/VisibleMarker.java
@@ -7,13 +7,13 @@ public class VisibleMarker {
 
     //What I need to create
     private int markerId;
-    private int xPos;
-    private int yPos;
+    private float xPos;
+    private float yPos;
     private ImageView image;
     private Bitmap solidColor;
     //TODO Fixa ett sätt att få referens från bilden jag valt för markören och spara den någonstans.
 
-    public VisibleMarker(int markerId, int xPos, int yPos){
+    public VisibleMarker(int markerId, float xPos, float yPos){
         this.markerId = markerId;
         this.xPos = xPos;
         this.yPos = yPos;
@@ -22,11 +22,11 @@ public class VisibleMarker {
     public int getMarkerId() {return markerId;}
     public void setMarkerId(int markerId) {this.markerId = markerId;}
 
-    public int getxPos() {return xPos;}
-    public void setxPos(int xPos) {this.xPos = xPos;}
+    public float getxPos() {return xPos;}
+    public void setxPos(float xPos) {this.xPos = xPos;}
 
-    public int getyPos() {return yPos;}
-    public void setyPos(int yPos) {this.yPos = yPos;}
+    public float getyPos() {return yPos;}
+    public void setyPos(float yPos) {this.yPos = yPos;}
 
     public ImageView getImage() {return image;}
     public void setImage(ImageView image) {
